requirements.txt
uagents==0.22.5
flask
openai
python-dotenv
